complete capabilities page
