+++ src/components/DataTable/DataTable.module.scss
@@ -1,0 +1,3 @@
+.tablePanel {
+  margin-top: 45px;
+}
