+++ src/components/DataTable/index.js
@@ -1,0 +1,3 @@
+import DataTable from "./DataTable"
+
+export { DataTable }
