+++ src/components/Home/index.js
@@ -1,0 +1,3 @@
+import Home from "./Home"
+
+export { Home }
