+++ src/components/Navbar/SideNavBar/index.js
@@ -1,0 +1,3 @@
+import SideNavBar from "./SideNavBar"
+
+export { SideNavBar }
