+++ src/pages/capabilities/SearchPanel/SearchPanel.module.scss
