+++ src/pages/capabilities/SearchPanel/index.js
