+++ src/pages/capabilities/capabilities.module.scss
