+++ src/pages/capabilities/index.js
@@ -1,0 +1,3 @@
+import Capabilities from "./capabilities"
+
+export { Capabilities }
