+++ src/sampleData/index.js
@@ -1,0 +1,3 @@
+import { capabilitiesData } from "./table.data"
+
+export { capabilitiesData }
