+++ src/sampleData/table.data.js
@@ -1,0 +1,48 @@
+export const capabilitiesData = {
+  columns: [
+    "x",
+    "IdPart",
+    "Part Number",
+    "UnPart Description",
+    "Base Part Number",
+    "Aircraft Series",
+    "Aircraft Model",
+    "Engine",
+    "ATA"
+  ],
+  rows: [
+    {
+      x: "1",
+      IdPart: "7BAA24F0-DFDA-41F7-B6CF-000B0709C882",
+      "Part Number": "642W9813-505",
+      "UnPart Description": "Wire Harness, Engine Control Unit to T/R",
+      "Base Part Number": "642W9813",
+      "Aircraft Series": "A318 - A321",
+      "Aircraft Model": "A318 - A321",
+      Engine: "CFM56-5",
+      ATA: "71"
+    },
+    {
+      x: "2",
+      IdPart: "7BAA24F0-DFDA-41F7-B6CF-000B0709C882",
+      "Part Number": "642W9813-505",
+      "UnPart Description": "Wire Harness, Engine Control Unit to T/R",
+      "Base Part Number": "642W9813",
+      "Aircraft Series": "A318 - A321",
+      "Aircraft Model": "A318 - A321",
+      Engine: "CFM56-5",
+      ATA: "71"
+    },
+    {
+      x: "3",
+      IdPart: "7BAA24F0-DFDA-41F7-B6CF-000B0709C882",
+      "Part Number": "642W9813-505",
+      "UnPart Description": "Wire Harness, Engine Control Unit to T/R",
+      "Base Part Number": "642W9813",
+      "Aircraft Series": "A318 - A321",
+      "Aircraft Model": "A318 - A321",
+      Engine: "CFM56-5",
+      ATA: "71"
+    }
+  ]
+}
